requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
openai==1.55.0
python-dotenv==1.0.1
python-multipart==0.0.12
supabase==2.18.1
pinecone==7.3.0
PyPDF2==3.0.1
tiktoken==0.11.0